restored utils package, moved non wallet functions from walletcore to utils and fixed other review change request
