+++ terminal/pages/helpers.go
@@ -5,12 +5,12 @@ import (
 	"math"
 
 	"github.com/decred/dcrd/dcrutil"
-	"github.com/raedahgroup/godcr/app/walletcore"
+	"github.com/raedahgroup/godcr/app/utils"
 )
 
 func maxDecimalPlaces(amounts []int64) (maxDecimalPlaces int) {
 	for _, amount := range amounts {
-		decimalPortion := walletcore.DecimalPortion(dcrutil.Amount(amount).ToCoin())
+		decimalPortion := utils.DecimalPortion(dcrutil.Amount(amount).ToCoin())
 		nDecimalPlaces := len(decimalPortion)
 		if nDecimalPlaces > maxDecimalPlaces {
 			maxDecimalPlaces = nDecimalPlaces
@@ -22,7 +22,7 @@ func maxDecimalPlaces(amounts []int64) (maxDecimalPlaces int) {
 func formatAmountDisplay(amount int64, maxDecimalPlaces int) string {
 	dcrAmount := dcrutil.Amount(amount).ToCoin()
 	wholeNumber := int(math.Floor(dcrAmount))
-	decimalPortion := walletcore.DecimalPortion(dcrAmount)
+	decimalPortion := utils.DecimalPortion(dcrAmount)
 
 	if len(decimalPortion) == 0 {
 		return fmt.Sprintf("%d%-*s DCR", wholeNumber, maxDecimalPlaces+1, decimalPortion)
